requirements.txt
pandas
jupyter
google-cloud-storage
google-cloud-bigquery
google-auth